refactor: move each stdlib function into its own module.
This commit refactors the standard library implementation such that each
function's implementation gets its own module.
